# docs/openapi/openapi_combined.yaml
openapi: 3.0.3
info:
  title: Social Media API
  description: Объединенная документация всех микросервисов
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: API Gateway (Development)
paths:
...
  /api/posts/{post_id}/stats:
    get:
      tags:
      - statistics
      summary: Get Post Stats
      description: Получение базовой статистики по посту
      operationId: get_post_stats_api_posts__post_id__stats_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostStatsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/posts/{post_id}/dynamics/views:
    get:
      tags:
      - statistics
      summary: Get Post Views Dynamics
      description: Получение динамики просмотров по посту
      operationId: get_post_views_dynamics_api_posts__post_id__dynamics_views_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          description: Количество дней для анализа
          default: 30
          title: Days
        description: Количество дней для анализа
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDynamicsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/posts/{post_id}/dynamics/likes:
    get:
      tags:
      - statistics
      summary: Get Post Likes Dynamics
      description: Получение динамики лайков по посту
      operationId: get_post_likes_dynamics_api_posts__post_id__dynamics_likes_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          description: Количество дней для анализа
          default: 30
          title: Days
        description: Количество дней для анализа
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDynamicsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/posts/{post_id}/dynamics/comments:
    get:
      tags:
      - statistics
      summary: Get Post Comments Dynamics
      description: Получение динамики комментариев по посту
      operationId: get_post_comments_dynamics_api_posts__post_id__dynamics_comments_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          description: Количество дней для анализа
          default: 30
          title: Days
        description: Количество дней для анализа
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDynamicsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/statistics/posts/top:
    get:
      tags:
      - statistics
      summary: Get Top Posts
      description: Получение топ постов по метрике
      operationId: get_top_posts_api_statistics_posts_top_get
      parameters:
      - name: metric
        in: query
        required: true
        schema:
          enum:
          - views
          - likes
          - comments
          type: string
          description: Метрика для топа
          title: Metric
        description: Метрика для топа
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          description: Количество постов в топе
          default: 10
          title: Limit
        description: Количество постов в топе
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopPostsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/statistics/users/top:
    get:
      tags:
      - statistics
      summary: Get Top Users
      description: Получение топ пользователей по метрике
      operationId: get_top_users_api_statistics_users_top_get
      parameters:
      - name: metric
        in: query
        required: true
        schema:
          enum:
          - views
          - likes
          - comments
          type: string
          description: Метрика для топа
          title: Metric
        description: Метрика для топа
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          description: Количество пользователей в топе
          default: 10
          title: Limit
        description: Количество пользователей в топе
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopUsersResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DynamicsPoint:
      properties:
        date:
          type: string
          title: Date
          description: Дата в формате YYYY-MM-DD
        count:
          type: integer
          minimum: 0.0
          title: Count
          description: Количество событий
      type: object
      required:
      - date
      - count
      title: DynamicsPoint
      description: Точка данных для динамики
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PostDynamicsResponse:
      properties:
        post_id:
          type: string
          title: Post Id
          description: ID поста
        metric:
          type: string
          enum:
          - views
          - likes
          - comments
          title: Metric
          description: Тип метрики
        dynamics:
          items:
            $ref: '#/components/schemas/DynamicsPoint'
          type: array
          title: Dynamics
          description: Данные динамики по дням
      type: object
      required:
      - post_id
      - metric
      - dynamics
      title: PostDynamicsResponse
      description: Схема ответа для динамики статистики поста
      example:
        dynamics:
        - count: 10
          date: '2024-01-01'
        - count: 15
          date: '2024-01-02'
        - count: 20
          date: '2024-01-03'
        metric: views
        post_id: '12345'
...
    PostStatsResponse:
      properties:
        post_id:
          type: string
          title: Post Id
          description: ID поста
        views_count:
          type: integer
          minimum: 0.0
          title: Views Count
          description: Количество просмотров
        likes_count:
          type: integer
          minimum: 0.0
          title: Likes Count
          description: Количество лайков
        comments_count:
          type: integer
          minimum: 0.0
          title: Comments Count
          description: Количество комментариев
      type: object
      required:
      - post_id
      - views_count
      - likes_count
      - comments_count
      title: PostStatsResponse
      description: Схема ответа для базовой статистики поста
      example:
        comments_count: 8
        likes_count: 25
        post_id: '12345'
        views_count: 150
...
    TopPost:
      properties:
        post_id:
          type: string
          title: Post Id
          description: ID поста
        count:
          type: integer
          minimum: 0.0
          title: Count
          description: Количество по метрике
      type: object
      required:
      - post_id
      - count
      title: TopPost
      description: Схема для топ поста
    TopPostsResponse:
      properties:
        metric:
          type: string
          enum:
          - views
          - likes
          - comments
          title: Metric
          description: Метрика сортировки
        posts:
          items:
            $ref: '#/components/schemas/TopPost'
          type: array
          title: Posts
          description: Список топ постов
      type: object
      required:
      - metric
      - posts
      title: TopPostsResponse
      description: Схема ответа для топ постов
      example:
        metric: views
        posts:
        - count: 1500
          post_id: '12345'
        - count: 1200
          post_id: '67890'
        - count: 800
          post_id: '11111'
    TopUser:
      properties:
        user_id:
          type: string
          title: User Id
          description: ID пользователя
        count:
          type: integer
          minimum: 0.0
          title: Count
          description: Количество по метрике
      type: object
      required:
      - user_id
      - count
      title: TopUser
      description: Схема для топ пользователя
    TopUsersResponse:
      properties:
        metric:
          type: string
          enum:
          - views
          - likes
          - comments
          title: Metric
          description: Метрика сортировки
        users:
          items:
            $ref: '#/components/schemas/TopUser'
          type: array
          title: Users
          description: Список топ пользователей
      type: object
      required:
      - metric
      - users
      title: TopUsersResponse
      description: Схема ответа для топ пользователей
      example:
        metric: likes
        users:
        - count: 250
          user_id: '1'
        - count: 180
          user_id: '2'
        - count: 150
          user_id: '3'
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...